+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+ORM M2 - 0.6.1.21
+------
+- Mise à jour de l'API MI pour les shares et vacation
+
+ORM M2 - 0.6.1.20
+------
+- Fix 0006229: [User Mél] Gérer le info=REPONSE:<adresse de reply-to> pour les listes
+
+ORM M2 - 0.6.1.19
+------
+- 0006229: [User Mél] Gérer le info=REPONSE:<adresse de reply-to> pour les listes
+
+ORM M2 - 0.6.1.18
+------
+- 0006225: [En attente] Un participant décliné ne doit pas avoir l'événement recréé
+
 ORM M2 - 0.6.1.17
 ------
 - Correctif double_authentification pour les Mel\User